condensing the classes and modules in the main file
