+++ bin/main.rb
@@ -1,9 +1,96 @@
 #!/usr/bin/env ruby
 # rubocop:disable Metrics/MethodLength
 require 'colorize'
-require_relative 'user'
-require_relative 'helpers'
-require_relative 'game_board'
+module AuthenticatingValues
+  SELECTION = %w[X O].freeze
+  NUMBERS_RANGES = (1..9).freeze
+
+  def self.validating_value(choice)
+    until SELECTION.any?(choice)
+      puts 'Please enter X or O'
+      choice = gets.chomp
+    end
+    choice
+  end
+
+  def self.numbers_validator(choice)
+    until NUMBERS_RANGES.any?(choice.to_i) || choice == 'q'
+      puts 'Please enter a number between 1 to 9 or q'
+      choice = gets.chomp
+    end
+    choice
+  end
+
+  def self.entry_space_validator(board, value)
+    option = false
+    until option
+      if value == board[value - 1]
+        option = true
+      else
+        puts 'That spot is taken chose again'.red
+        new_input = gets.chomp
+        value = new_input.to_i
+        option = false
+      end
+    end
+    value
+  end
+
+  def self.authenticating_winner(board, player)
+    board.winner.each do |item|
+      next unless board.board[item[0]] == player.value &&
+                  board.board[item[1]] == player.value &&
+                  board.board[item[2]] == player.value
+
+      puts "Congratulatios #{player.name} you Won!!".green
+      return true
+    end
+    false
+  end
+end
+
+class Board
+  attr_accessor :board
+  attr_reader :winner
+
+  def initialize
+    @board = [1, 2, 3, 4, 5, 6, 7, 8, 9]
+    @winner = [
+      [0, 1, 2],
+      [3, 4, 5],
+      [6, 7, 8],
+      [0, 3, 6],
+      [1, 4, 7],
+      [2, 5, 8],
+      [0, 4, 8],
+      [2, 4, 6]
+    ]
+  end
+
+  def display_board
+    puts " \t\t \n".yellow
+    puts " \t\t#{@board[0]}|#{@board[1]}|#{@board[2]}"
+    puts " \t\t----- ".yellow
+    puts " \t\t#{@board[3]}|#{@board[4]}|#{@board[5]}"
+    puts " \t\t----- ".yellow
+    puts " \t\t#{@board[6]}|#{@board[7]}|#{@board[8]}"
+    puts " \t\t----- \n".yellow
+  end
+end
+
+class Player
+  attr_reader :name, :value
+
+  def initialize(name, value)
+    @name = name
+    @value = value
+  end
+
+  def show_players_info
+    puts "player #{@name.green} choose #{@value.yellow}"
+  end
+end
+
 class GameUi
   attr_reader :title, :instructions
